Moving container component into sub-folder
