+++ src/App.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import App from './App';
-import Container from './components/Container';
+import Container from './components/container/Container';
 
 describe('the App component', () => {
   it('renders', () => {
